refactor: improve modularity and docs for textarea
